+++ frameworks/solid/src/primitives/index.ts
@@ -1,3 +1,4 @@
 export * from './createForm/index.ts';
 export * from './useField/index.ts';
 export * from './useFieldArray/index.ts';
+export * from './useFormData/index.ts';